requirements.txt
requests
websocket-client
colorama
python-dotenv